Sunday work

+++ PRGReaderLibrary/Constants.cs
@@ -36,16 +36,16 @@ namespace PRGReaderLibrary
 
     public static class TypesConstants
     {
-        public const int LOCAL_VARIABLE = 0x82;
-        public const int FLOAT_TYPE = 0x83;
-        public const int LONG_TYPE = 0x84;
-        public const int INTEGER_TYPE = 0x85;
-        public const int BYTE_TYPE = 0x86;
-        public const int STRING_TYPE = 0x87;
-        public const int FLOAT_TYPE_ARRAY = 0x88;
-        public const int LONG_TYPE_ARRAY = 0x89;
-        public const int INTEGER_TYPE_ARRAY = 0x8A;
-        public const int BYTE_TYPE_ARRAY = 0x8B;
-        public const int STRING_TYPE_ARRAY = 0x8C;
+        public const byte LOCAL_VARIABLE = 0x82;
+        public const byte FLOAT_TYPE = 0x83;
+        public const byte LONG_TYPE = 0x84;
+        public const byte INTEGER_TYPE = 0x85;
+        public const byte BYTE_TYPE = 0x86;
+        public const byte STRING_TYPE = 0x87;
+        public const byte FLOAT_TYPE_ARRAY = 0x88;
+        public const byte LONG_TYPE_ARRAY = 0x89;
+        public const byte INTEGER_TYPE_ARRAY = 0x8A;
+        public const byte BYTE_TYPE_ARRAY = 0x8B;
+        public const byte STRING_TYPE_ARRAY = 0x8C;
     }
 }
